refactor: 중복 컴포넌트 정리 (Header, MainLoading, ProgressBar)
- shared/ui/header/Header.tsx 삭제 (shared/ui/Header.tsx로 통합)
- shared/ui/loading/MainLoading.tsx 삭제 (shared/ui/MainLoading.tsx로 통합, fixed 포지셔닝)
- shared/ui/progress/ProgressBar.tsx 삭제 (dead code)
- login layout, ProtectedRoute import 배럴 경유로 변경
