+++ omechu-app/src/shared/ui/MainLoading.tsx
@@ -1,4 +1,4 @@
-export const MainLoading = () => {
+export function MainLoading() {
   return (
     <>
       <style>
@@ -12,19 +12,16 @@ export const MainLoading = () => {
           }
         `}
       </style>
+
       <div
-        className="absolute top-0 left-0 h-full w-full bg-cover bg-center bg-no-repeat"
-        style={{
-          backgroundImage: "url('/image/loading-bg.png')",
-        }}
+        className="fixed inset-0 z-[10000] bg-cover bg-center bg-no-repeat"
+        style={{ backgroundImage: "url('/image/loading-bg.png')" }}
       >
         <div
           className="animate-food-cycle absolute top-1/2 left-1/2 h-[170px] w-[175px] -translate-x-1/2 -translate-y-1/2 bg-[url('/image/food-sprite.svg')]"
-          style={{
-            backgroundSize: "525px 170px",
-          }}
-        ></div>
+          style={{ backgroundSize: "525px 170px" }}
+        />
       </div>
     </>
   );
-};
+}
